requirements.txt
python-magic
python-magic-bin
libmagic
docopt
typing
PyYAML
Pillow
eyed3
six
